% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_layer_ble}
\alias{add_layer_ble}
\title{Add a layer to bin label explore plot}
\usage{
add_layer_ble(
  conn,
  signal_code,
  procedure,
  color,
  fgnames = c(),
  inst_source = "AFSC"
)
}
\arguments{
\item{conn}{The database connection}

\item{signal_code}{integer id for signal, ie 1 (RW), 2 (GS). see db}

\item{procedure}{integer id for procedure. see db}

\item{color}{color. doesn't scale with counts, just presence.}

\item{fgnames}{optional fgs to consider.}

\item{inst_source}{AFSC if you don't want to include other lab data in the db}
}
\value{
ggplot layer to + to bin label explore ggplot object
}
\description{
Add a layer to a bin label explore plot. No easy way to include legend, use for quick comparisons.
}
