% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_negatives}
\alias{bin_negatives}
\title{produce bin negatives}
\usage{
bin_negatives(
  data,
  FG,
  bintype,
  analyst = "previous",
  procedure = NULL,
  signal_code = NULL,
  format = "db"
)
}
\arguments{
\item{data}{The detection data}

\item{FG}{The filegroup}

\item{bintype}{"LOW","REG", or "SHI" . Will automatically break bins so that it is on best timescale for queries.}

\item{procedure}{procedure id to populate in output object. defaults to procedure in 1st row if provided in data.}

\item{signal_code}{signal_code to populate in output object. defaults to signal_code in 1st row if provided in data.}

\item{format}{detx or db format}
}
\value{
data frame of bin negatives.
}
\description{
create data frame of bin negatives for a detector deployment on arbitrary fg
}
