% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{detx_pgpamdb_det_rough_convert}
\alias{detx_pgpamdb_det_rough_convert}
\title{convert a dataset from detx format to db format}
\usage{
detx_pgpamdb_det_rough_convert(
  conn,
  detx_data,
  procedure,
  strength,
  depnames = NULL
)
}
\arguments{
\item{conn}{The database connection}

\item{detx_data}{The dataset in detx format}

\item{procedure}{procedure id for the output detections (no equivalent in detx)}

\item{strength}{strength_codes id for the output detections (no equivalent in detx)}

\item{depnames}{additional column to provide in case of file name abiguity. Must match to values in data_collection 'name' column.}
}
\value{
a db format data frame
}
\description{
convert a dataset from detx format to db format. Could use work to be a little more general in cases where detx has variable fields.
}
