% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{detx_to_db}
\alias{detx_to_db}
\title{Convert between INSTINCT detx detection format and database detection format}
\usage{
detx_to_db(conn, dataset)
}
\arguments{
\item{conn}{The database connection}

\item{dataset}{The R detx object.}
}
\value{
result
}
\description{
Convert between INSTINCT detx detection format and database detection format. Assumes postgresql db with existing connection.
}
