% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fg_breakbins}
\alias{fg_breakbins}
\title{break FG into n sized bins}
\usage{
fg_breakbins(data, interval, format = "db")
}
\arguments{
\item{data}{The filegroup data.}

\item{interval}{size in seconds for max bins.}

\item{format}{detx or db format}
}
\value{
FG with rows greater than interval broken into interval sized segments.
}
\description{
return a vers of a filegroup which has segments broken into sections of size interval.
}
