% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_proj_functions.R
\name{FN_validate}
\alias{FN_validate}
\title{Check FW labels}
\usage{
FN_validate(conn, calcdata, original_folder)
}
\arguments{
\item{conn}{The database connection}

\item{calcdata}{data output from format_FW}

\item{original_folder}{the folder name containing original labels (output folder for FinReview.R)}
}
\value{
a table which shows original and back calculated labels side by side
}
\description{
Check FW labels against original labels
}
