% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_proj_functions.R
\name{format_FW}
\alias{format_FW}
\title{Format FW data on ANALYSIS}
\usage{
format_FW(conn, mooringname)
}
\arguments{
\item{conn}{The database connection}

\item{mooringname}{The name of the dataset as on ANALYSIS. Can be new naming or old naming convention (must match data_collection.name or data_collection.historic_name)}
}
\value{
a dataframe ready to be loaded into db.
}
\description{
Format the FW data on ANALYSIS into a db ready format
}
