% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{i_neg_interpolate}
\alias{i_neg_interpolate}
\title{Create assumed detections for implied negative data}
\usage{
i_neg_interpolate(data, FG, high_freq, procedure, signal_code, analyst)
}
\arguments{
\item{data}{The dataset. Uses start and end time and start and end file. column names need to be in db format. must correspond with fg.}

\item{FG}{The filegroup, in db format (bins.soundfiles_id, bins.seg_start,bins.seg_end, soundfiles.datetime)}

\item{high_freq}{high frequency of the implied negative detection.}

\item{procedure}{procedure id for the implied negative detections}

\item{signal_code}{signal id for the implied negative detections}

\item{analyst}{analyst id for the implied negative detections}
}
\value{
dataframe in db format of implied negative detections
}
\description{
Create assumed detections from ground truth detections.
}
