% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lookup_from_match}
\alias{lookup_from_match}
\title{Look up database table by named column values and return id}
\usage{
lookup_from_match(conn, tablename, vector, match_col, idname = "id")
}
\arguments{
\item{conn}{The database connection}

\item{tablename}{The name of the data table in the database}

\item{vector}{The values to compare with database table.}

\item{match_col}{The name of the database column to compare the vector columns to}

\item{idname}{character string specifying name of primary key.}
}
\value{
lookup table of ids and the initial vector values
}
\description{
Look up database table by named column values and return id. Useful for data conversions. Assumes postgresql db with existing connection.
}
