% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{table_dataset_lookup}
\alias{table_dataset_lookup}
\title{Query with table of values}
\usage{
table_dataset_lookup(
  conn,
  from_db_query,
  dataset,
  data_types,
  interpret_as = NULL
)
}
\arguments{
\item{conn}{The database connection}

\item{from_db_query}{The query, including select and joins, which will be joined to user R dataset.}

\item{dataset}{The R object to update, all columns must have equivalents in database table. Defaults to column names.}

\item{data_types}{The postgresql target column data types. Must be provided and explicit.}

\item{interpret_as}{Ordered vector which maps dataset columns to database (including join syntax, if required) columns}
}
\value{
a data frame corresponding to returned columns specified in 'from_db_query' argument
}
\description{
Query a user selected table or join of tables, using a dataset from R session. interpret as maps column names of R dataset, in
order, to names of equivalent columns on database.
}
