% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{table_delete}
\alias{table_delete}
\title{Delete data in the database by id}
\usage{
table_delete(conn, tablename, ids, idname = "id", hard_delete = FALSE)
}
\arguments{
\item{conn}{The database connection}

\item{tablename}{The name of the data table in the database}

\item{ids}{The id vector to delete. All rows with id in vector will be deleted.}

\item{idname}{character string specifying name of primary key.}

\item{hard_delete}{bool specifying whether to 'hard delete', refering to the archiving behavior of detections table on first delete. Only does anything if tablename = 'detections'}
}
\value{
result
}
\description{
Delete data in a named database table. Assumes postgresql db with existing connection.
}
