% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{table_insert}
\alias{table_insert}
\title{Insert data into the database}
\usage{
table_insert(conn, tablename, dataset, colvector = NULL, idname = "id")
}
\arguments{
\item{conn}{The database connection}

\item{tablename}{The name of the data table in the database}

\item{dataset}{The R object to update- must have id column matching type of tablename, and all columns must have equivalents in database table.}

\item{colvector}{Vector which specifies specific columns from dataset to update: otherwise assumes all matching column names from dataset to be updated}

\item{idname}{character string specifying name of primary key.}
}
\value{
a data frame of the ids of the uploaded dataset, in order, from the target table.
}
\description{
Insert data into a named database table. Assumes postgresql db with existing connection.
}
