% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_label_explore}
\alias{bin_label_explore}
\title{Plot all instances of a label code}
\usage{
bin_label_explore(
  conn,
  bin_label,
  inst_source = "AFSC",
  plot_sds = 4,
  bin_type = 1
)
}
\arguments{
\item{conn}{The database connection}

\item{bin_label}{bin label code, ie 'RW'}

\item{inst_source}{AFSC if you don't want to include other lab data in the db}

\item{plot_sds}{standard deviations for the color scale. Smaller = more similar colors, larger = more dynamic colors.}

\item{bin_type}{integer id for bin type- 1 = LOW, 2 = REG, 3 = SHI. Affects n but shouldn't affect plot too much.}
}
\value{
ggplot object
}
\description{
Plot all instances of a label code. Shows spread of label across data regardless of analysis. Gray signifies data present, no
effort, boxes around tile indicate at least one positive presence instance.
}
