% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pamdbConnect}
\alias{pamdbConnect}
\title{Pamdb connect}
\usage{
pamdbConnect(dbname, infoScript, sslKey_path, sslCert_path)
}
\arguments{
\item{dbname}{The name of the db}

\item{infoScript}{The script that when run, populates dsn_ variables in local function}

\item{sslKey_path}{The path to the sslKey}

\item{sslCert_path}{The path to the sslCert}
}
\value{
The database connection
}
\description{
Connect to the pamdb database, using locally available files.
}
